requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.46
cryptography
mysqlclient
Flask-Migrate
